Klimakoffer.jl API

Klimakoffer._calc_insolation — Method

calcinsolation() auxiliar function of calcsolarforcing

Klimakoffer.area_weighted_average — Method

Computes mean temperature in the globe at a specific time

Klimakoffer.calc_diffusion_coefficients — Function

calcdiffusioncoefficients() Calculate the diffusion coefficients at finest grid level.

Klimakoffer.calc_radiative_cooling_co2 — Function

calcradiativecooling_co2() Computes the CO2 parameter depending on the CO2 concentration

Default CO2 concentration is 315 ppm (equivalent to year 1950)

Klimakoffer.calc_solar_forcing — Function

calcsolarforcing()

Default s0 is 1371.685 W/m²

Default orbital parameters of correspond to year 1950 AD: ecc = 0.016740 ob = 0.409253 per = 1.783037

Klimakoffer.compute_equilibrium! — Method

compute_equilibrium!(...)

rel_error is the tolerance for global temperature equilibrium (default is 2e-5).

max_years is the maximum number of annual cycles to be computed when searching for equilibrium

Klimakoffer.compute_evolution! — Method

compute_evolution!(...)

Compute the evolution of the mean temperature with varying CO2 levels.